Committing lots of changes

+++ generate_gif_q_learning.py
@@ -4,21 +4,27 @@ from os import listdir, unlink
 from os.path import isfile, join
 from lib.utils import *
 from lib.generate_graphics import *
+# from lib import plotting
 from datetime import datetime
+from q_learning import *
+from cat_mouse_env import CatMouseEnv_binary_reward, CatMouseEnv_proximity_reward
+from importlib import reload
 
-# This
+# import lib.generate_graphics
+# reload(lib.generate_graphics)
 
-
-def play_cat_and_mouse(board_height, board_width, show_figs = True):
+def play_cat_and_mouse_q_learning(cat_policy, board_height, board_width, show_figs = True):
     # assumimg cat has sight 2 in each direction (i.e. can see a 5x5 grid around iteself)
     # cat and mouse move uniformly (can move any direction, or stay put, with prob 1/9)
-    # cat policy doesn't update - stays uniform
     # note that if either cat or mouse attempts to move into the wall (even diagonally) they stay where they are
+
+    if not len(cat_policy) == board_height * board_width * board_height * board_width:
+        raise Exception('Policy incompatible with given board size')
 
     start_time = datetime.now().strftime('%Y%m%d_%H%M')
 
-    for file in listdir('gif_graphics'):
-        file_with_path = join('gif_graphics', file)
+    for file in listdir('graphics_gif'):
+        file_with_path = join('graphics_gif', file)
         if isfile(file_with_path):
             unlink(file_with_path)
 
@@ -34,18 +40,21 @@ def play_cat_and_mouse(board_height, board_width, show_figs = True):
     board[cat_pos] = 'C'
     board[mouse_pos] = 'M'
 
-    mouse_pos_prob_dist = initialise_mouse_prob_dist(board_height, board_width, cat_pos, mouse_pos)
+    # mouse_pos_prob_dist = initialise_mouse_prob_dist(board_height, board_width, cat_pos, mouse_pos)
+    mouse_pos_prob_dist = np.zeros((board_height, board_width))
+    mouse_pos_prob_dist[mouse_pos] = 1
 
     print('Starting position')
 
-    filename = filename = 'gif_graphic_' + start_time + '_0' + '.png'
+    filename = 'gif_graphic_' + start_time + '_0' + '.png'
     filenames.append(filename)
     generate_fig(cat_pos, mouse_pos, mouse_pos_prob_dist, board_height, board_width, filename, show_figs)
 
     iter = 1
     while cat_pos != mouse_pos:
-        cat_vert_move = np.random.choice((-1,0,1))
-        cat_horz_move = np.random.choice((-1,0,1))
+        state_index = positions_to_state_index(cat_pos, mouse_pos, board_height, board_width)
+        cat_action_index = cat_policy[state_index]
+        cat_vert_move, cat_horz_move = action_index_to_moves(cat_action_index)
         mouse_vert_move = np.random.choice((-1,0,1))
         mouse_horz_move = np.random.choice((-1,0,1))
         new_cat_pos = (cat_pos[0] + cat_vert_move, cat_pos[1] + cat_horz_move)
@@ -58,38 +67,42 @@ def play_cat_and_mouse(board_height, board_width, show_figs = True):
         board[cat_pos] = 'C'
         board[mouse_pos] = 'M'
 
-        if cat_can_see_mouse(cat_pos, mouse_pos):
-            mouse_pos_prob_dist = np.zeros((board_height, board_width))
-            mouse_pos_prob_dist[mouse_pos] = 1
-        else:
-            new_mouse_pos_prob_dist = np.zeros((board_height, board_width))
-            for row in range(board_height):
-                for col in range(board_width):
-                    for row_offset in [-1, 0, 1]:
-                        for col_offset in [-1, 0, 1]:
-                            if (((row + row_offset) in range(board_height)) and ((col + col_offset) in range(board_width))):
-                                if not ((abs(row + row_offset - cat_pos[0]) <= 2) and (abs(col + col_offset - cat_pos[1]) <= 2)):
-                                    new_mouse_pos_prob_dist[row + row_offset, col + col_offset] += (1/9) * mouse_pos_prob_dist[row,col]
-                            else:
-                                if not ((abs(row - cat_pos[0]) <= 2) and (abs(col - cat_pos[1]) <= 2)):
-                                    new_mouse_pos_prob_dist[row, col] += (1/9) * mouse_pos_prob_dist[row,col]
-
-            sum_mouse_pos_prob_dist = np.sum(new_mouse_pos_prob_dist)
-            new_mouse_pos_prob_dist /= sum_mouse_pos_prob_dist
-            mouse_pos_prob_dist = deepcopy(new_mouse_pos_prob_dist)
-
-
+        # fully observable so cat can always see mouse
+        mouse_pos_prob_dist = np.zeros((board_height, board_width))
+        mouse_pos_prob_dist[mouse_pos] = 1
 
         print('\n Iteration %d' % iter)
 
         filename = 'gif_graphic_' + start_time + '_' + str(iter) + '.png'
         filenames.append(filename)
 
-        gif_filename = 'gif_graphics/the_gif/output_' + start_time + '.gif'
+        gif_filename = 'graphics_gif/the_gif/output_' + start_time + '.gif'
 
         generate_fig(cat_pos, mouse_pos, mouse_pos_prob_dist, board_height, board_width, filename, show_figs)
         iter += 1
 
     generate_gif(filenames, gif_filename)
 
-play_cat_and_mouse(4,6)
+board_height = 4
+board_width = 6
+
+Q_binary_reward, stats_binary_reward = q_learning(CatMouseEnv_binary_reward(board_height, board_width), 100000)
+plot_episode_stats(stats_binary_reward, 1000)
+
+
+cat_policy_binary_reward = get_greedy_policy(Q_binary_reward)
+
+play_cat_and_mouse_q_learning(cat_policy_binary_reward, board_height, board_width)
+
+show_policy(cat_policy_binary_reward, (1,3))
+
+
+Q_proximity_reward, stats_proximity_reward = q_learning(CatMouseEnv_proximity_reward(board_height, board_width), 100000)
+plot_episode_stats(stats_proximity_reward, 1000)
+
+
+cat_policy_proximity_reward = get_greedy_policy(Q_proximity_reward)
+
+play_cat_and_mouse_q_learning(cat_policy_proximity_reward, board_height, board_width)
+
+show_policy(cat_policy_proximity_reward, (0,5))
